added classes for generating pojos from KustoResultSetTable instances (#159)
* added classes for generating pojos from KustoResultSetTable instances

This commit includes three new classes:
1. KustoType which is a generic class for representing different Kusto
column types and bind them to different Java types

2. KustoColumn which represents a Kusto column - it has a name,
nullability information, and a type (KustoType)

3. KustoResultMapper which can be set up to parse KustoResultSetTable
instances and return a list of pojos. The setup phase is done using a
builder that takes the pojo constructor, as well as zero or more
instances of KustoColumn along with setter functions to set pojo's
fields based on the column values.

This mechanism can potentially be extended to use reflection based on
the target pojo class instead of the manual KustoResultMapper setup

* Update data/src/main/java/com/microsoft/azure/kusto/data/KustoType.java

Changed documentation link to international version

Co-authored-by: ohad bitton <32278684+ohadbitt@users.noreply.github.com>

* Use String.format() for exception messages

* Introduced ObjectPopulator class, ordinal column mapping

- Unified KustoColumn and QueryResultColumnInfo into the new
ObjectPopulator class
- Added support for manually specified column ordinals
- extractList finds column indexes once per result set to save on column
lookups

* KustoResultMapper extractSingle - don't create empty objects

If the KustoResultSetTable is null, the proper response is to return a
null object, not an unpopulated one

* ObjectPopulator - always use column index, formatting

populateFrom used to get objects from the ResultSet by the column name
unless an ordinal was set. Now it always uses the ordinal (looking it up
if necessary).

* KustoType formatting - don't align field declarations

* KustoResultMapperTest

The test uses a mocked KustoResultMapper to test the pojo mapping
functionality of the KustoResultMapper

* Update KustoType.java

* Removed addNonNullableColumn/addNonNullableColumn, added javadoc

Instead of individual addNonNullableColumn/addNullableColumn methods,
exposed the more generic addColumn methods. Additionally, added javadoc
to document the addColumn parameters

* Renamed ObjectPopulator to KustoResultColumnPopulator

* KustoResultColumnPopulator::populateFrom use passed ordinal

populateFrom used to always lookup the column ordinal by name instead of
using the passed ordinal. Now it's always using the ordinal, moving the
responsibility of correctly determining that ordinal to the caller.

* renamed variable in KustoResultColumnPopulator::populateFrom

* KustoResultMapper::extractList renamed variable

* added "single" to names of tests of KustoResultMapper::extractSingle

* Javadoc fixes in KustoType and KustoResultMapper

Co-authored-by: ohad bitton <32278684+ohadbitt@users.noreply.github.com>