Support AAD Managed Identity authentication (#194)
* Support AAD Managed Identity authentication

* Update data/src/main/java/com/microsoft/azure/kusto/data/auth/ManagedIdentityTokenProvider.java

Co-authored-by: AsafMah <asafmahlev@microsoft.com>

* Fix scope

* Fix space

* Downgrade azure-identity package to avoid showstopper bug causing ManagedIdentity auth to fail.

Note that there still remains 1 change to be made in this PR - to set the scope based on the Kusto Service Resource ID from CloudInfo. In support of this, I have a 22-file change to broadly improve how the TokenProviders consume from CloudInfo, but am hesitating to push that before understanding the initial CloudInfo design decisions.

At a high level, I would like to change the throws list when creating a client. It currently throws URISyntaxException, because it only uses CloudInfo when a token is requested. However, I don't see why we don't make use of CloudInfo when creating the client, in which case we should return a DataClientException/DataServiceException. This is a breaking change and a non-trivial paradigm shift, so it's worth discussing first.

I will begin discussing this with Asaf tomorrow.

* Resolve dependency conflicts

* Reorganize CloudInfo code

* Fix compilatino error

* Remove superfluous method (determineAadAuthorityUrl)

* Remove duplicate line

* Fix cloudinfo initialization

* Fix E2E tests (have been broken for a while)

* Switch order of clientUrl to be first

Co-authored-by: AsafMah <asafmahlev@microsoft.com>